[http clients code] Refactor common code into common.go
Changelog: None
